+++ juke_auth/urls.py
@@ -1,6 +1,6 @@
 from django.urls import include, path
 from rest_framework import routers
-from app import views
+from juke_auth import views
 
 router = routers.DefaultRouter()
 router.register(r'users', views.UserViewSet)
