+++ juke_auth/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.models import User, Group
 from rest_framework import viewsets
 from rest_framework import permissions
-from app.serializers import UserSerializer
+from juke_auth.serializers import UserSerializer
 
 
 class UserViewSet(viewsets.ModelViewSet):
