+++ juke_backend/settings.py
@@ -46,7 +46,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'django_filters',
     'django_extensions',
-    'app',
+    'juke_auth',
 ]
 
 MIDDLEWARE = [
